Move the helper classes into their own files
