+++ Vic2ToHoI4/Source/V2World/Issues/PartiesIssueHelper.cpp
@@ -1,0 +1,13 @@
+#include "PartiesIssueHelper.h"
+
+
+
+Vic2::PartiesIssueHelper::PartiesIssueHelper(std::istream& theStream)
+{
+	registerRegex("[A-Za-z_]+", [this](const std::string& theIssue, std::istream& theStream) {
+		IssueHelper helper(theStream);
+		issues.push_back(helper);
+	});
+
+	parseStream(theStream);
+}
