+++ Vic2ToHoI4Tests/HoI4WorldTests/States/HoI4StateTests.cpp
@@ -7,7 +7,7 @@
 #include "../Vic2ToHoI4/Source/HOI4World/Map/CoastalProvinces.h"
 #include "../Vic2ToHoI4/Source/HOI4World/States/HoI4State.h"
 #include "../Vic2ToHoI4/Source/OutHoi4/States/OutHoI4State.h"
-#include "../Vic2ToHoI4/Source/V2World/Issues.h"
+#include "../Vic2ToHoI4/Source/V2World/Issues/Issues.h"
 #include "../Vic2ToHoI4/Source/V2World/Province.h"
 #include "gtest/gtest.h"
 #include <optional>
