+++ Vic2ToHoI4Tests/Vic2WorldTests/ProvinceTests.cpp
@@ -1,4 +1,4 @@
-#include "../Vic2ToHoI4/Source/V2World/Issues.h"
+#include "../Vic2ToHoI4/Source/V2World/Issues/Issues.h"
 #include "../Vic2ToHoI4/Source/V2World/Province.h"
 #include "gtest/gtest.h"
 #include <sstream>
